Enable single tap rebinding for hwkey back and home
Initially, this policy was put in place to safeguard
the UX by preventing a usecase where the device could
be put in a state without back or home available. However,
if a user puts the device in this state, it is easy enough
to navigate to button settings to fix it.

Change-Id: I6b71f9e54fcb99356e07f236e9ed926ac5fc017e